Build and release actions (#1)
This change adds two GitHub Action workflows for building, testing and releasing
the controller. The CI workflow is triggered on all pushes and the Release flow
is only triggered on new Git tag pushes. Use the release make target to create
these tags, as this will create a new commit and update the deploy/operator.yaml
specification.

Further some common targets are added to a Makefile to ease development.