Implement simple PostgreSQLUser CRD (#4)
This change implements the basic role creation in PostreSQL based on a
PostgreSQLUser custom resource. It also adds a number of make targets for
testing the controller.

It handles detection of resources and creation of the role based on its name. It
is possible to configure a role prefix and the associated roles to grant the
created one e.g. rds_iam.