Implement read and write fields (#9)
This change adds the read and write fields for PostgreSQLUser custom resources.
The code handles granting of privileges in PostgreSQL based on the requirements
defined in custom resources.

When a new CR is reconciled a new role is created as before, but now grants are
added to database/schemas according to the read and write specification.

All the SQL logic in package postgresqluser is moved into package postgres to
increase readability and seperation of concerns. The integration tests also
depend on package postgresqldatabase which is avoided with this as well.

The postgres docker image used in the test deployment is changed to
timm/postgres-logging. The only difference is that query logging is enabled on
that image.