Handle multiple hosts for database controller
A PostgreSQLDatabase resource specifies what host it requests a database on but
the field was not read.

This changes respects the field by adding a map of host names to credentials.
For each CR a lookup to this map is made to find the proper credentials for
interacting with the host.