Configuration flags (#12)
This change adds controller configuration with flags.

The configuration fields of both PostgreSQLUser and PostgreSQLDatabase are now
populated in a parseFlags function.

The two flagsets are registered in main and allows each controller to specify
its own flags.

One caveat about this solution is that overlapping flags are ignored. Package
pflag does not allow overlapping flags so host-credentials are duplicated for
each controller (host-credentials-user and host-credentials-database) even
though they are almost always expected to be identical.

I would have like to inject all this configuration from main, but the design of
operator-sdk to rely on globals for controller registrations makes injection of
the values very challenging.