Add hostname to postgres.Database function
When a database is created a schema with the same name is created with the newly
created service role. A bug in this handling would try to connect to localhost
when creating the schema. This works in the integration tests where the database
is located on localhost but in a cluster this fails.

This change adds a host parameter to the Database function ensuring that we
connect to the right host when connecting with the service role.