Harden kube functions (#21)
This change hardens the error handling scenarios of the kube functions.

Function ResourceValue now verifies the existence of value before accessing
resource KeySelector structs. This fixes a panic when the resources reference
ConfigMaps as the secret value is evaluated first and non-existent.

A guard on non-existing keys is added as well so custom resources referencing
unknown fields of a config map or secret will be detected.

Lastly I renamed a test case that was no longer correct after moving the
functionality out of package postgresqldatabase.