Integration tests in CI (#27)
This change adds make targets for running a PostgreSQL instance in Docker.

When running the ci GitHub Action workflow we run make test/integration
instead of test/unit that will start the PostgreSQL instance before running
the go test command.

We need to change the assertion in one of the tests as the error message is
different in the action environment than on a local development machine. In the
action a TCP error like this is returned: dial tcp: lookup unknown: Temporary failure in name resolution where we on development machines get dial tcp: lookup unknown: no such host. It is not important for the test case so now we
assert on a dial tcp error and ignore the underlying reason.