Make the daemon loop work
Implements a synchronous sync strategy inspired by how flux handles git and job
syncs.

A syncSoon buffered channel is used to limit the number of simultanious syncs.
A timer requests syncs to this channel regularly.

A hook for completions is used for instrumentation probes to track the duration
of syncs in a histogram.