Move access granting into package grants (#38)
This change moves all grouping and resolving of user grants into a grants
package. This will make it reusable and make the currently large controller
implementation more manageable.

This is done in preparation of expiration logic that will need much of the same
logic.

In general the methods are just moved into a new package and references are
updated. No logic changes inside the methods.