Daemon implementation for scheduled tasks (#39)
This change adds a daemon implementation responsible for running a function in a
scheduled manner.

It will be used to implement expiration handling of grants. This is added in a
future pull request.

The main idea is to have a way of setting up a recursing function and take som
action on it without having to think about Go routines etc. in the domain logic.

An instrumentation package is added as well with a single sync histogram to track
sync durations.