Reduce the number of open database connections
To reduce the number of open connections from the controller we reduce the
number of idle connections to zero and max open connections to 1. This will
result in a lot of connections comming and going during resource sync loops, but
will ensure that we drop the connections as soon as possible.

Currently connections can stay in the connection pool for a long period of time
and when reconciling all resources in the cluster the connections will stay
longer than we want.

As each resource uses a new sql.DB as we don't know if the next reconsiliation
is against the same database we do not facilitate the build in connection
pooling of sql.DB.

There is also added a missing Close call on the service connection when creating
new databases.