Fix database connection usage (#44)
To reduce the number of open connections from the controller we reduce the
number of idle connections to zero. This will result in a lot of connections
coming and going during resource sync loops, but will ensure that we drop the
connections as soon as possible.

Currently connections can stay in the connection pool for a long period of time
and when reconciling all resources in the cluster the connections will stay
longer than we want.

As each resource uses a new sql.DB as we don't know if the next reconciliation
is against the same database we do not facilitate the build in connection
pooling of sql.DB.

Further more a resync-period flag that controls the before mentioned resync loop
as to let us test this behaviour without having to rebuild the controller.

There is also added a missing Close call on the service connection when creating
new databases.