Log when allDatabases results on no grants
If flag allDatabases is set on an access request we silently proceed without
granting anything.

This is expected, but can be hard to understand from the logs.

This change adds a log line when this happens so one can see that we correctly
handled the request but found nothing to add.