Support shared database (#48)
This change set introduces support for shared databases between different users.
This requires the addition of two fields to the PostgreSQLDatabase resource, ie.
User and IsShared.

By specifying the User name to use for creating the service roles, we support
multiple users on the same database with different passwords. It will fallback
to Name if it is empty to support existing resources.

The IsShared boolean instructs the controller to not grant ownership of the
database to the reconciled user, as this would effectively revoke ownership from
another users database.