Grant access based on schema name
When using shared databases the database name is the same between multiple
schemas, but when granting access from a PostgreSQLUser we request access on a
specific schema of a shared database. This results in read and readwrite access
grants are based on the database name roles intead of the schema specific ones.
This results in the grants not actually granting access to the schemas.

This change set changes the grants to based on the schema name instead and thus
the read and readwrite roles are properly granted.

It also fixes the limitation of allDatabases that would not work with shared
databases by using the User as the schema name.