Log reconciliation errors with request id
If a reconciliation fails the controller-runtime controller will log the error,
but doing so without the request id.

This change set adds another error log to the controllers on exit logging the
same error but with the contextual logger as to get the request id logged along
with the error.