Support missing user for AllDatabases grants
If a PostgreSQLUser requests access to all databases with the AllDatabases flag
and one of the found PostgreSQLDatabase resources has not set the User field an
"no value" error is returned and granting fails.

This change set adds a fallback in case of the missing User field to the
database name as we do in package postgres when setting up database.