Add support for failures for allDatabases requests (#53)
If a database resources is not configured correctly, ie. unknown config map
reference or is temporarily in a bad state the controller is not able to grant
access to any databases when a user requests it with the allDatabases field.

This fails to support practical usage where some database objects will be in a
bad state at any time and thus puts the controller out of operation.

This change will only log an error if a database object cannot be resolved but
will continue reconciling any databases that are valid.

A controller test is added with a fake kubernetes client that contains two
databases, one good and one referencing an unknown config map. When reconciling
a PostgreSQLUser with allDatabases set we expect the reconciliation to complete
for the valid database.