Expire roles when access requests expires (#54)
This PR adds handling of the Start and Stop fields in PostgreSQLUser
resources. It also moves much of the granting logic out of
postgresqluser_controller.go and into a grants/sync.go file where it fits
better with all the database logic.

Overall the main difference is that we now inspect existing roles before
granting the required ones. We then diff the existing ones with the ones we
expected and removes those that are no longer expected. As a safe guard only
roles ending with _read and _readwrite are touched for now. These are the
roles we grant.

Expiration in general is handled in the groupAccesses method of
grants.Granter where Start and Stop are inspected and the specifications
are filtered out if they are not relevant. Combined with the described role
removal above, this leads roles to expire.

Note that we do not have a daemon implementation triggering this reconciliation.
We use operator-sdk's sync interval (default 10 hours) to control how often this
should be done. If we need to be more aggressive with revoking access on
expiration the --resync-period can be set accordingly.