Use request logger for grants.Granter
Currently we store a general logger in grants.Granter. Any log lines does that
way not have any context fields from a caller, e.g request ids.

This change set uses a logger in the first argument instead (like other methods
in the controller) and passes that along. This also fixes a panic when running
the controller as grants.Granter.Log was not set during startup.