Grant roles across hosts instead of databases (#59)
Currently roles are granted on a per databases basis. The role resolvement logic
can there for only know the existing of roles applicable to that database and
thus revokes access to any other database on the same host.

This change set groups the requests by host (instance) instead of host/database
and thus lets the role resolvement logic span across all databases on the host.
This reduces the number of connections to each host, as we only need one instead
of one per database, and also reduces the number of calls to the database as we,
in the same fashion, only need a single GRANT and REVOKE query per host. It also
fixes a bug where roles just granted were revoked shortly after when requesting
access to multiple databases on the same host.

This implementation change is only around the grouping mechanism and with this
simplification removes the ugly host/database concatenation that was spread out
through package grants.