Add support for DROP and ALTER (#61)
Currently it is not possible to DROP and ALTER tables with a valid write
permission. In some situations this is impractical, eg. fast development
iterations in an environment.

This change set introduces a new 'extended' field to write access
specifications. This instructs to controller to make the requesting user able to
DROP and ALTER tables.

It is not possible to grant these privileges directly on a role, so instead the
controller grants the service user role to the developer making the developer
inherit the owner privileges of objects created by the service user. This makes
it possible to use DROP and ALTER.

A new 'readowningwrite' role is introduced to the managed set of roles and this
is the one added to developer roles at run-time when ever they request extended
write access. The feature is guarded by flag 'extended-write-enabled' and
disabled by default.

Two other bugs were found and fixed during development of this feature.

Detection of removable roles was buggy, but functional, as the existing manged
roles had the same prefix, ie. read and readwrite. This is changed to match the
complete role suffix.

The SyncUser method would never grant write access if there was not also a read
access request. The bug was in a fast-return path of groupAccesses that only
considered read specs. This is changed to consider both read and write specs.