Ignore non-running databases when reconciling user
Currently, if a database is in an invalid of failed reconciliation phase,
PostgreSQLUser resources cannot be reconciled if they enable flag allDatabases.
This results in all PostgreSQLUser resources with this flag will never be
applied for all other databases available as well.

This change introduces a guard in the allDatabases grouping skipping any
non-running databases. If a database is under startup, ie. empty phase, they are
also ignored but will be considered on the next reconciliation.