Add support for params on host credentials
Currently the controller defaults to sslmode=disable for all host connections.
Further more additional connection paramters cannot be added if one would like
to do so.

This change introduces the option for adding params=<connection-params> to host
credentils in the --host-crendetial flag to allow specifying custom paramters.

If no params are specified, it defaults to existing behaviour, that is setting
sslmod=disable.