Fix RBAC on configmaps and secrets
The controller needs access to secrets and config maps if resource references
are used by PostgreSQLDatabase or PostgreSQLUser resources. Currently if any of
these are used and the deafult role is used, the controller is unable to
reconsile.

This change adds the kubebuilder annotations to ensure the role is correct.