Add PostgreSQLHostCredentials custom resource (#90)
Add a new PostgreSQLHostCredentials Custom Resource that replaces the
host-credentials controller flag.

The new resource will be used to find admin credentials for instances by using a
new reference field in the PostgreSQLDatabase resource. The resource collects
credentials in a structure way opposed to the current flag and makes it possible
to reference other Kubernetes resources for providing values for eg. passwords.

The motivation for adding the resource is around the last feature. When using
the controller along with operators like Crossplane.io or Kubegres the instance
credentials are provided through Kubernetes resources like Secrets and
ConfigMaps. With the new resource it now becomes possible to reference these. It
also avoids the odd workflow of adding new instances and having to update the
controllers flags to make it possible to work with them.

The change also adds context propagation to some controller methods where it was
missing.

Closes #82 and leaves #91 for later implementation.

Co-authored-by: Bjørn <bjoern.soerensen@gmail.com>