Fix bug with unknown PostgreSQLHostCredentials reference (#93)
Currently if a PostgreSQLDatabase references an unknown
PostgreSQLHostCredentials resource reconciliation is disabled. This is not
intended because the referenced resources might show up in the near future, eg.
it has to be unsealed. The bug is due to unknown references leading to an
errors.Invalid error which skips future reconciliations.

This change includes two integration tests of a PostgreSQLHostCredentials
reference in a PostgreSQLDatabase resource which validates the bug and fix. The
change also adds some additional logging to give insights into the decision
process along with cleaning up the error handling of said logic.

Lastly the Scheme field of the PostgreSQLDatabaseReconciler struct is removed as
it is unused.