Fix panic on PostgreSQLDatabase resource delete
Currently if a PostgreSQLDatabase resource is removed the
PostgreSQLDatabaseReconciler.reconcile method will return early on the Not Found
error from the Kubernetes API. The PostgreSQLDatabaseReconciler.Reconcile method
will after this try to update the status of the resource which panics as it was
never found on the status value.

This change adds a guard on the status.update method which will make it safe to
do on non-existing resources and effectively make the update a noop. This fixes
the panic and ensures that the controller handles the deletion successfully.