Support dot characters in PostgreSQLUser spec.name (#95)
Currently the spec.name field is mapped to both the AWS user email and
postgresql role name. This model is not correct as emails might contain "."
(dot) characters but these are not allowed in PostgreSQL roles.

This change updates our model to separate these concerns. A sanitized version of
the field is then used for the role while the original field is used as the
email.