Upgrade Go and dependencies (#96)
This change upgrades all direct dependencies where applicable along with the Go
runtime.

The controller-runtime contains several major changes including the upgrade of
logr which contains several breaking changes. Good thing is that the new stuff
makes a lot of our logging types redundant so they are all removed.

The GitHub actions were running an old Go version which is no longer supported
by controller-runtime so we had to upgrade those as well.