Update IAM role when reconciling
Currently if a user exists in the IAM policy the controller does not update any
of the properties for it. This means that the policies become stale when making
changes to how roles are managed, eg. when
dd8fbee449da2725f19f32b55df6aa14e69e4c04 introduced changes to how role names
are created.

This change renames the iam.AddUser function to iam.EnsureUser while updating
its implementation to always ensure that the policy is as it should be.