Add proper requeue strategy for database resources
Currently if an error happens during reconciliation of a PostgreSQLDatabase
resource we log an error log with the error. This is noisy and not really an
error in the controller but wrong usage from the client.

Further more we never retry any errors automatically before an update is caused
which makes turn around time for eg. a secret appearing longer.

This change updates the Reconcile method to select a proper strategy based on
errors to know when to retry. It also stops returning errors to the controller
manager which avoids error logs from it. Overall we get reduced error logs and
better reconciliation handling.