feat: add service user crd
This is an addition to support the use-case where we may want a user spanning a database host. The user should be automatically created, with a given password and associated roles.

The user is granted basic settings, such as connect, password etc, such that it can operate as a service user, and access the database itself, with no permissions.

This feature is proposed because we've got a need for easily creating and managing roles across many databases with the same settings and behavior

Signed-off-by: Kasper J. Hermansen <contact@kjuulh.io>